All processes excepted GradientDynamics Behaviorally tested. Separated Initialization tests to test_process.py and behavioral tests to test_models.py
